OK, let's commit this as v1.
It isn't really finished, but I need Entry for LDAP::Dist, so I need to
be able to package this now. I've left out the dep on BMORROW::Util,
because I don't want to package that yet. (I'll have to see where I get
to with mkbootstrap: probably most of ::NSS will move into -Dist
anyway.)